Enhance ValidateField method and update version
Updated the `ValidateField` method in the `FluentValidation<TModel>` class to include an optional `highlightInvalidField` parameter, which defaults to `true`. This change allows for optional highlighting of invalid fields during validation and adds a notification for validation state changes.

Incremented the version number in `NuvTools.Validation.AspNetCore.Blazor.csproj` from `9.2.0` to `9.2.1` to reflect the new release.